refactor: cleanup testing of tree and also fix bug
